+++ djpcms/templatetags/djpmodutils.py
@@ -2,6 +2,7 @@ from django import template
 from django.db import models
 
 from djpcms.utils import force_unicode, smart_str
+from djpcms.views import appsite
 
 register = template.Library()
 
@@ -65,3 +66,11 @@ def objvalue(obj,name):
         value = getattr(obj, name)
     return value
     
+    
+@register.filter
+def objtable(obj):
+    appmodel = appsite.site.for_model(obj.__class__)
+    if appmodel:
+        return appmodel.opts.totable(obj)
+    else:
+        return obj